class01/pyproject.toml
[project]
name = "classai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.16",
    "python-decouple>=3.8",
]
